Renames validate-input to validateInput.
