+++ lib/Model.js
@@ -29,7 +29,7 @@ var isJSONEnvelope = require("./support/isJSONEnvelope");
 var isJSONGraphEnvelope = require("./support/isJSONGraphEnvelope");
 var jsong = require("falcor-json-graph");
 var ID = 0;
-var validateInput = require("./support/validate-input");
+var validateInput = require("./support/validateInput");
 var noOp = function() {};
 var getCache = require("./get/getCache");
 var get = require("./get");
