+++ lib/support/validateInput.js
@@ -1,7 +1,7 @@
 var isArray = Array.isArray;
-var isPathValue = require("./is-path-value");
-var isJsonGraphEnvelope = require("./is-json-graph-envelope");
-var isJsonEnvelope = require("./is-json-envelope");
+var isPathValue = require("./../support/isPathValue");
+var isJSONGraphEnvelope = require("./../support/isJSONGraphEnvelope");
+var isJSONEnvelope = require("./../support/isJSONEnvelope");
 var pathSyntax = require("falcor-path-syntax");
 
 /**
@@ -32,12 +32,12 @@ module.exports = function validateInput(args, allowedInput, method) {
         }
 
         // jsonGraph {jsonGraph: { ... }, paths: [ ... ]}
-        else if (isJsonGraphEnvelope(arg) && allowedInput.jsonGraph) {
+        else if (isJSONGraphEnvelope(arg) && allowedInput.jsonGraph) {
             valid = true;
         }
 
         // json env {json: {...}}
-        else if (isJsonEnvelope(arg) && allowedInput.json) {
+        else if (isJSONEnvelope(arg) && allowedInput.json) {
             valid = true;
         }
